db-sharding/requirements.txt
Flask>=3.1.0,<4.0.0
psycopg2-binary>=2.9.10,<3.0.0
